extract LoginRouter to production code
